Add fstab and helpers
- Add a fstab for centos-7 & fedora-latest
- Remove matchers
-
