--- .foodcritic
@@ -1,1 +1,0 @@
-~FC016
