+++ .travis.yml
@@ -40,7 +40,7 @@ env:
   - INSTANCE=remove-ubuntu-1604 CHEF_VERSION=12.7.2
   - INSTANCE=persist-amazonlinux
   - INSTANCE=persist-centos-6
-    #- INSTANCE=persist-centos-7 MISSING FSTAB FILE
+  - INSTANCE=persist-centos-7
   - INSTANCE=persist-debian-8
   - INSTANCE=persist-debian-9
   - INSTANCE=persist-ubuntu-1404
